Moved structs and interfaces to the appropriate places
