+++ pkg/framework/plugins/requesthandling/basemodelextractor/adapters_store.go
